pyproject.toml
# note: this file is for local development, currently the file is not used to build
# the plugin
[tool.poetry]
name = "basic-games"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psutil = "^5.9"
vdf = "3.4"
lzokay = "1.1.5"
PyCryptoDome = "^3.20"
joblib = "^1.4"

[tool.poetry.group.dev.dependencies]
mobase-stubs = { version = "^2.5.0.dev15", allow-prereleases = true }
pyqt6 = "^6.6.1"
pyright = "^1.1.350"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
ruff = "^0.2.1"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^24.2.0"
flake8-black = "^0.3.6"
types-psutil = "^5.9.5.20240205"
poethepoet = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
lint-black = "black --check --diff ."
lint-isort = "isort -c ."
lint-mypy = "mypy ."
lint-ruff = "ruff ."
lint-pyright = "pyright ."
lint-all.sequence = [
    "lint-black",
    "lint-isort",
    "lint-mypy",
    "lint-ruff",
    "lint-pyright",
]
lint-all.ignore_fail = "return_non_zero"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

[tool.isort]
profile = "black"
multi_line_output = 3
skip_gitignore = true

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
exclude = ["lib", "games/quarantine"]
platform = "win32"

[[tool.mypy.overrides]]
module = "vdf.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lzokay.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pyright]
exclude = ["lib", "**/.*", "venv"]
typeCheckingMode = "strict"
pythonPlatform = "Windows"
reportMissingModuleSource = false
